+++ src/core/webdriver.py
@@ -1,5 +1,9 @@
+from typing import Tuple
+
 import chromedriver_autoinstaller
 import undetected_chromedriver as uc
+from selenium.webdriver.support import expected_conditions as EC
+from selenium.webdriver.support.ui import WebDriverWait
 from seleniumwire.undetected_chromedriver.v2 import Chrome, ChromeOptions
 
 
@@ -27,3 +31,17 @@ def init_seleniumwire_webdriver(*args):
     chrome_options.add_argument("--disable-software-rasterizer")
     chrome_options.add_argument("--disable-blink-features=AutomationControlled")
     return Chrome(options=chrome_options)
+
+
+def wait_for_presence_of_element(
+    driver, locator: Tuple[str, str], timeout: int = 30
+) -> None:
+    """
+    Wait for the presence of a particular web element. This does not necessarily mean that the element is visible.
+
+    :param driver: Selenium WebDriver instance
+    :param locator: Locator for the element.
+    :param timeout: Maximum wait time in seconds
+    :raises: TimeoutException if element is not found within timeout
+    """
+    return WebDriverWait(driver, timeout).until(EC.presence_of_element_located(locator))
